+++ js/previewplugin.js
@@ -53,10 +53,18 @@
                 id: 'Files_3dViewer'
             });
 
+            var spinner = function () {
+                return OC.imagePath('files_3dviewer', 'loading-spinner.gif');
+            }
+
+            // console.log(spinner);
+
             var $viewerOverlay = $('<div id="viewer_overlay"></div>');
             var $viewerContainer = $('<div id="viewer_container"></div>');
             var $canvasContainer = $('<div id="canvas_container"></div>');
-            var $loadingOverlay = $('<div id="loading_overlay"><img src="' + OC.imagePath('core', 'loading.gif') + '"></div>');
+            // var $loadingOverlay = $('<div id="loading_overlay"><img src="' + OC.imagePath('core', 'loading.gif') + '"></div>');
+            // var $loadingOverlay = $('<div id="loading_overlay"><img src="../files_3dviewer/img/loading_spinner.gif"></div>');
+            var $loadingOverlay = $('<div id="loading_overlay"><img src="' + OC.imagePath('files_3dviewer', 'loading_spinner.gif') + '"></div>');
 
             $viewerContainer.append($canvasContainer);
             $viewerContainer.append($loadingOverlay);
@@ -75,29 +83,42 @@
 
             var element = document.getElementById("canvas_container");
 
-            var camera_state={
+            var camera_state = {
                 "position": {"x": 100, "y": 100, "z": 74.39999999999999},
                 "up": {"x": 100, "y": 1, "z": 1000},
                 "target": {"x": 100, "y": 100, "z": 100}
             };
-            var stlViewer = new StlViewer(
+
+            var init_state = {
+                "position": {"x": 0, "y": 0, "z": 0},
+                "up": {"x": 0, "y": 0, "z": 0},
+                "target": {"x": 0, "y": 0, "z": 0}
+            };
+
+            var stl_viewer = new StlViewer(
                 element,
                 {
                     models: [
                         {
                             id: 1,
-                            filename: "/remote.php/webdav" + dir + "/" + filename
+                            filename: "/remote.php/webdav" + dir + "/" + filename,
+                            view_edges: false,
+                            opacity: 0.9,
+                            display: "flat",
+                            color: "#d7d7d7",
+                            animation: {
+                                delta: {
+                                    rotationx: 0, rotationy: 0, rotationz: 1, msec: 1000, loop: true
+                                },
+                            }
                         }
                     ],
                     // jszip_path:"jszip.min.js",
                     // jszip_utils_path:"jszip-utils.min.js",
                     // load_three_files: "/",
                     loading_progress_callback: this._loadingProgress,
-                    // set_camera_state: {"position":{"x":100,"y":100,"z":74.39999999999999},"up":{"x":100,"y":1,"z":1000},"target":{"x":100,"y":100,"z":100}},
+                    all_loaded_callback: this.init_orientation
                 });
-
-            // var_dump(stlViewer.camera);
-            // this.init_orientation(stlViewer)
         },
 
         _loadingProgress: function (load_status, load_session) {
@@ -124,37 +145,38 @@
         },
 
         init_orientation: function (stl_viewer) {
+            var last_camera_state = this.get_camera_state();
 
-        var last_camera_state = stl_viewer.get_camera_state();
-        // if (!last_camera_state) return;
-        var s='';
+            if (!last_camera_state) return;
 
-        //rotate camera (whole scene)
-        var distance=Math.max(Math.abs(stl_viewer.camera.position.x),Math.abs(stl_viewer.camera.position.y),Math.abs(stl_viewer.camera.position.z));
+            var distance = Math.max(Math.abs(this.camera.position.x), Math.abs(this.camera.position.y), Math.abs(this.camera.position.z));
 
-        stl_viewer.set_camera_state({position:{x:distance,y:distance,z:distance},up:{x:0,y:1,z:0},target:last_camera_state.target});
-        scene_rotation=s;
+            this.set_camera_state({
+                position: {x: distance, y: distance, z: distance},
+                up: {x: 0, y: 0, z: 1},
+                target: last_camera_state.target
+            });
 
-
-        $id('orientation_icon').setAttribute('class', 'huge vs-font vs-'+s);
-        $id('orientation_icon_menu').setAttribute('class', 'huge vs-font vs-'+s);
-    },
+            this.set_rotation(1, Math.PI / 10, -(Math.PI / 10), Math.PI, 0);
+        },
 
         /**
          * Load the editor control bar
          */
         loadControls: function (file, context) {
             var html =
-                '<small class="filename">' + escapeHTML(file) + '</small>'
-                + '<small class="unsaved-star" style="display: none">*</small>'
-                + '<small class="saving-message">'
-                + '</small>'
-                + '<button id="viewer_close" class="icon-close svg"></button>';
+                '<small class="filename">' + escapeHTML(file) + '</small>';
 
             var controlBar = $('<div id="viewer_controls"></div>').html(html);
 
             var $progressBarTotal = $('<div id="progress_total"><progress id="pbtotal" class="p-total" value="0" max="1"></progress></div>');
 
+            var $closeBtn = $('<button/>', {
+                id: 'viewer_close',
+                class: 'icon-close svg',
+            })
+
+            controlBar.append($closeBtn);
             controlBar.append($progressBarTotal);
 
             $('#viewer_container').append(controlBar);